opscopilot-devui-demo/requirements.txt
agent-framework --pre
agent-framework-devui --pre
python-dotenv
nest_asyncio
fastapi
uvicorn
